02:34 | 02:34 AM | Sun | Nov | 05/11/2023 0.0.19 create Array.every.lua create Array.filter.lua create Array.find.lua create Array.findIndex.lua create Array.includes.lua create Array.map.lua create Array.reduce.lua create Array.some.lua create operator_nullish_coalescing.lua create operator_optional_chaining.lua create operator_spread.lua create operator_ternary.lua create factorial.lua create fizzbuzz.lua create functions.lua create Array.js create Array.py create Array.php create Array.go create Array.pl create Array.jl create Array.lua

+++ codes/julia/factorial.jl
@@ -3,7 +3,7 @@ using Statistics
 
 println("\n# Factorial(n) in Julia")
 
-function factorial_v1(number)
+function factorial_v1(number::Int)
     if ((number > 1) === false)
         return 1
     end
@@ -14,7 +14,7 @@ function factorial_v1(number)
     return result
 end
 
-function factorial_v2(number)
+function factorial_v2(number::Int)
     if ((number > 1) === false)
         return 1
     end
@@ -30,7 +30,7 @@ function factorial_v2(number)
     return result
 end
 
-function factorial_v3(number)
+function factorial_v3(number::Int)
     if ((number > 1) === false)
         return 1
     end
